Remove gpio-keys for camera_snapshot and focus
We expose gpio 67 and 68 to userspace in the PIO HAL, which allows users
to access these pins from applications. Since these are mapped to mic
mute and camera mute, we expect applications to attach InputDrivers to
these pins. To resolve this, we remove these references so that these
pins are not considered claimed by the gpio driver.

Bug: 72185990
Test: Pins are now exportable from sysfs and test applications work.
Change-Id: Ib3464726685f2392576654cf3536f31085b9ca33