Switch APQ8053 Lite Dragon to gpio-based vol keys
Now that the RESIN interrupt can be mapped to a GPIO, we establish that
mapping here, and also remove the gpio-keys entry for volume up. With
this change, buttons will not generate key events, and applications will
be expected to read the state of the pin through the framework.

Bug: 72956949
Test: Verified that volume keys are accessible over GPIO and callbacks
fire correctly upon registration of EDGE_BOTH.
Change-Id: I2aa57a833cab1183330af9fe1f42c1a6ae7c0f31