This is a cherry-pick of a fix from b/73597588.
Comments from original commit:

qcacld-2.0: Fix check for num_ac in iface stats event handler

In function wma_unified_link_iface_stats_event_handler, num_ac
is received from the firmware and is used in the loop to populate
values into result. However the memory for result  is allocated
for WIFI_AC_MAX (which is 4). While copying the data from fw, the
fw loop is looped from i = 0 to i < num_ac which is 0 to 3.
So the value of num_ac = 4 from the fw is not an issue.

Modify the upper limit check for num_ac to be greater than
WIFI_AC_MAX instead of greater than equal to WIFI_AC_MAX.

Bug:
https://buganizer.corp.google.com/issues/79214920

Change-Id: Id9145a4b934b6a0f2ca38193357c34e4b769edd3