Revert "ARM: msm: dts: Increase sdio clock as 100Mhz on FEP"
This reverts commit 8202473e1ff0b0bb7590bb5be8512994ea1142fe.

The suspicion is that this change has introduced instability into the
SDIO card. With the limited logs we have, we clearly see the bus failing
and thus, the card stopping entirely. Since this is a critical failure
that would prevent updates unless the user explicitly rebooted, we
revert this as a stop-gap measure to fall back to a known working speed.

Bug: 79923387
Test: Stability tests have been running successfully for an hour. This
option has been set to 50MHz for the greater part of field testing, and
no errors have been reported.

Change-Id: Ic00c1883f95132b89c1eb369ff99d87137187ffa